Set up integration test for creating screenshots.
In preparation for better testing/play store entry.
