+++ test_driver/integration_test.dart
@@ -1,0 +1,18 @@
+import 'dart:io';
+import 'package:integration_test/integration_test_driver_extended.dart';
+
+Future<void> main() async {
+  await integrationDriver(
+    onScreenshot: (
+      String screenshotName,
+      List<int> screenshotBytes, [
+      Map<String, Object?>? args,
+    ]) async {
+      final File image = await File(
+        'screenshots/$screenshotName.png',
+      ).create(recursive: true);
+      image.writeAsBytesSync(screenshotBytes);
+      return true;
+    },
+  );
+}
